+++ korgw/src/commonMain/kotlin/com/soywiz/kgl/KmlGlProxy.kt
@@ -998,7 +998,7 @@ open class KmlGlProxy(val parent: KmlGl) : KmlGl() {
 		after("vertexAttrib4fv", sparams, "$res")
 		return res
 	}
-	override fun vertexAttribPointer(index: Int, size: Int, type: Int, normalized: Boolean, stride: Int, pointer: Int): Unit {
+	override fun vertexAttribPointer(index: Int, size: Int, type: Int, normalized: Boolean, stride: Int, pointer: Long): Unit {
 		val sparams = "$index, $size, $type, $normalized, $stride, $pointer"
 		before("vertexAttribPointer", sparams)
 		val res = parent.vertexAttribPointer(index, size, type, normalized, stride, pointer)
@@ -1439,7 +1439,7 @@ open class KmlGlFastProxy(val parent: KmlGl) : KmlGl() {
 	override fun vertexAttrib4fv(index: Int, v: FBuffer): Unit {
 		return parent.vertexAttrib4fv(index, v)
 	}
-	override fun vertexAttribPointer(index: Int, size: Int, type: Int, normalized: Boolean, stride: Int, pointer: Int): Unit {
+	override fun vertexAttribPointer(index: Int, size: Int, type: Int, normalized: Boolean, stride: Int, pointer: Long): Unit {
 		return parent.vertexAttribPointer(index, size, type, normalized, stride, pointer)
 	}
 	override fun viewport(x: Int, y: Int, width: Int, height: Int): Unit {
